+++ admin/tabs/logging/LoggingTab.js
@@ -14,7 +14,7 @@ export const loggingTab = hoistCmp.factory(
     () => tabContainer({
         model: {
             route: 'default.logging',
-            switcherPosition: 'left',
+            switcher: {orientation: 'left'},
             tabs: [
                 {id: 'viewer', icon: Icon.fileText(), content: logViewer, omit: !XH.getConf('xhEnableLogViewer', true)},
                 {id: 'config', icon: Icon.settings(), content: logLevelPanel}