+++ cmp/tab/TabContainer.js
@@ -7,7 +7,7 @@
 import {hoistCmp, ModelPublishMode, uses, XH} from '@xh/hoist/core';
 import {tabContainerImpl as desktopTabContainerImpl} from '@xh/hoist/dynamics/desktop';
 import {tabContainerImpl as mobileTabContainerImpl} from '@xh/hoist/dynamics/mobile';
-import {throwIf} from '@xh/hoist/utils/js';
+import {apiRemoved} from '@xh/hoist/utils/js';
 import PT from 'prop-types';
 import {TabContainerModel} from './TabContainerModel';
 
@@ -15,10 +15,10 @@ import {TabContainerModel} from './TabContainerModel';
  * Display a set of child Tabs and (optionally) a switcher control.
  *
  * By default this TabContainer will install a TabSwitcher above the Tabs to control the currently
- * displayed Tab. The 'TabContainerModel.switcherPosition' property can be adjusted to place
- * the switcher control on alternative edges of the container.
+ * displayed Tab. The 'TabContainerModel.switcher' property can be adjusted to place the switcher
+ * control on alternative edges of the container.
  *
- * If the switcherPosition is set to 'none' then no TabSwitcher will be installed.  This latter case
+ * If `switcher` is set to false then no TabSwitcher will be installed.  This setting
  * is useful for applications that wish to place an associated TabSwitcher elsewhere in the graphical
  * hierarchy (e.g. a shared menu bar), or control the visible Tab directly via other means.
  *
@@ -33,14 +33,11 @@ export const [TabContainer, tabContainer] = hoistCmp.withFactory({
     className: 'xh-tab-container',
 
     render(props) {
-        throwIf(
-            props.switcherPosition,
-            "'switcherPosition' is no longer present on TabContainer.  Please specify on TabContainerModel instead."
-        );
-
+        apiRemoved(props.switcherPosition, 'switcherPosition', 'Please specify TabContainerModel.switcher instead.');
         return XH.isMobileApp ? mobileTabContainerImpl(props) : desktopTabContainerImpl(props);
     }
 });
+
 TabContainer.propTypes = {
     model: PT.oneOfType([PT.instanceOf(TabContainerModel), PT.object])
 };