+++ cmp/tab/TabContainerModel.js
@@ -6,7 +6,7 @@
  */
 import {HoistModel, managed, PersistenceProvider, RefreshMode, RenderMode, XH} from '@xh/hoist/core';
 import {action, observable} from '@xh/hoist/mobx';
-import {ensureUniqueBy, throwIf} from '@xh/hoist/utils/js';
+import {ensureUniqueBy, apiRemoved, throwIf} from '@xh/hoist/utils/js';
 import {find, isUndefined, without, difference} from 'lodash';
 import {TabModel} from './TabModel';
 
@@ -32,8 +32,8 @@ export class TabContainerModel {
     /** @member {string} */
     @observable activeTabId;
 
-    /** @member {string} */
-    switcherPosition;
+    /** @member {boolean|Object} */
+    switcher;
 
     /** @member {boolean} */
     track;
@@ -55,8 +55,9 @@ export class TabContainerModel {
      * @param {?string} [c.route] - base route name for this container. If set, this container will
      *      be route-enabled, with the route for each tab being "[route]/[tab.id]".
      *      Cannot be used with `persistWith`.
-     * @param {string} [c.switcherPosition] - Position of the switcher docked within this component.
-     *      Valid values are 'top', 'bottom', 'left', 'right', or 'none' if no switcher shown.
+     * @param {(boolean|Object)} [c.switcher] - indicates whether to include a default switcher
+     *      docked within this component. Specify as a boolean or an object containing props for a
+     *      TabSwitcher component. Set to false to not include a switcher. Defaults to true.
      * @param {boolean} [c.track] - True to enable activity tracking of tab views (default false).
      * @param {RenderMode} [c.renderMode] - strategy for rendering child tabs. Can be set
      *      per-tab via `TabModel.renderMode`. See enum for description of supported modes.
@@ -71,17 +72,25 @@ export class TabContainerModel {
         tabs = [],
         defaultTabId = null,
         route = null,
-        switcherPosition = XH.isMobileApp ? 'bottom' : 'top',
+        switcher = true,
         track = false,
         renderMode = RenderMode.LAZY,
         refreshMode = RefreshMode.ON_SHOW_LAZY,
         persistWith,
-        emptyText = 'No tabs to display.'
+        emptyText = 'No tabs to display.',
+        switcherPosition
     }) {
-        throwIf(!['top', 'bottom', 'left', 'right', 'none'].includes(switcherPosition), 'Unsupported value for switcherPosition.');
+        apiRemoved(switcherPosition, 'switcherPosition', 'Please specify `switcher` instead.');
+
+        // Create default switcher props
+        if (switcher === true) {
+            switcher = {orientation: XH.isMobileApp ? 'bottom' : 'top'};
+        }
+
+        throwIf(switcher && !['top', 'bottom', 'left', 'right'].includes(switcher.orientation), 'Unsupported value for switcher orientation.');
         throwIf(route && persistWith, '"persistWith" and "route" cannot both be specified.');
 
-        this.switcherPosition = switcherPosition;
+        this.switcher = switcher;
         this.renderMode = renderMode;
         this.refreshMode = refreshMode;
         this.defaultTabId = defaultTabId;